+++ src/events/events.lua
@@ -8,6 +8,7 @@ local E = Addon:GetModule("Events")
 local events = {
   "LootReceived",
   "MoneyReceived",
+  "SavedVariablesLoaded",
   "TexturedLootMessage"
 }