+++ src/saved-variables.lua
@@ -11,7 +11,10 @@ local GLOBAL_SV = "__LOOTIFICATIONS_ADDON_GLOBAL_SAVED_VARIABLES__"
 
 local function globalDefaults()
   return {
-    moneyNotifications = true
+    anchorPoint = {},
+    maxNotifications = Addon.MAX_NOTIFICATIONS_DEFAULT,
+    moneyNotifications = true,
+    notificationFadeOutDelay = Addon.NOTIFICATION_FADE_OUT_DELAY_DEFAULT
   }
 end
 
@@ -37,6 +40,10 @@ local function depopulate(t, defaults)
   end
 end
 
+local function clamp(value, minValue, maxValue)
+  return math.min(math.max(value, minValue), maxValue)
+end
+
 -- ============================================================================
 -- Events
 -- ============================================================================
@@ -47,9 +54,17 @@ EventManager:Once(E.Wow.PlayerLogin, function()
   populate(_G[GLOBAL_SV], globalDefaults())
   local global = _G[GLOBAL_SV]
 
+  -- Clamp `maxNotifications` between min and max values.
+  global.maxNotifications = clamp(global.maxNotifications, Addon.MAX_NOTIFICATIONS_MIN, Addon.MAX_NOTIFICATIONS_MAX)
+  -- Clamp `notificationFadeOutDelay` between min and max values.
+  global.notificationFadeOutDelay = clamp(global.notificationFadeOutDelay, Addon.NOTIFICATION_FADE_OUT_DELAY_MIN,
+    Addon.NOTIFICATION_FADE_OUT_DELAY_MAX)
+
   function SavedVariables:Get()
     return global
   end
+
+  EventManager:Fire(E.SavedVariablesLoaded)
 end)
 
 -- Deinitialize.