+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.2.0] - 2023-12-31
+
+## Added
+
+- Command: `/lootifications anchor` to toggle the notifications anchor
+- Command: `/lootifications anchor reset` to reset the notifications anchor
+- Command: `/lootifications delay <integer>` to set the delay time before notifications fade out
+- Command: `/lootifications max <integer>` to set the maximum number of displayed notifications
+
 ## [1.1.1] - 2023-12-28
 
 ### Changed